// frontend/src/app/api/model/dashboardLearning.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DashboardLearning { 
    cardsTotal: number;
    inProgressCount: number;
    milestonesTotal: number;
    milestonesDone: number;
    milestonesCompletedLast7Days: number;
}